requirements/base.txt
# Backend Core
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.3

# Database
psycopg2-binary==2.9.9
pymongo==4.8.0

# Cache and Redis
redis==5.0.7
django-redis==5.4.0

# Async Tasks
celery==5.3.7
celery[redis]==5.3.7
flower==2.0.1

# Security
django-axes==6.4.0
django-ratelimit==4.1.0
django-otp==1.5.0
django-oauth-toolkit==1.7.1
django-csp==3.8
django-cryptography==1.1
bleach==6.1.0
python-clamd==3.2.0
django-graphql-jwt==0.4.0

# Search
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0

# Monitoring and Logging
sentry-sdk==2.8.0
django-prometheus==2.3.1
structlog==24.2.0

# APIs
graphene-django==3.2.2
channels==4.1.0
channels-redis==4.2.0
requests==2.32.3

# Environment
python-decouple==3.8

# File handling
Pillow==10.4.0
python-magic==0.4.27
openpyxl==3.1.5

# Utilities
ipython==8.26.0
